This adds some benchmarks copied from jech/samplebuilder.
```
name                           old time/op    new time/op    delta
SampleBuilderSequential-8         313ns ± 6%     235ns ±10%  -24.93%  (p=0.029 n=4+4)
SampleBuilderLoss-8               296ns ± 6%     193ns ±11%  -34.64%  (p=0.029 n=4+4)
SampleBuilderReordered-8          338ns ± 5%     241ns ±16%  -28.66%  (p=0.029 n=4+4)
SampleBuilderFragmented-8         299ns ± 4%     230ns ±14%  -23.12%  (p=0.029 n=4+4)
SampleBuilderFragmentedLoss-8     281ns ± 5%     190ns ±19%  -32.63%  (p=0.029 n=4+4)

name                           old alloc/op   new alloc/op   delta
SampleBuilderSequential-8          336B ± 0%      336B ± 0%     ~     (all equal)
SampleBuilderLoss-8                300B ± 0%      290B ± 0%   -3.33%  (p=0.029 n=4+4)
SampleBuilderReordered-8           335B ± 0%      336B ± 0%   +0.30%  (p=0.029 n=4+4)
SampleBuilderFragmented-8          336B ± 0%      336B ± 0%     ~     (all equal)
SampleBuilderFragmentedLoss-8      295B ± 0%      280B ± 0%   -5.08%  (p=0.029 n=4+4)

name                           old allocs/op  new allocs/op  delta
SampleBuilderSequential-8          4.00 ± 0%      4.00 ± 0%     ~     (all equal)
SampleBuilderLoss-8                3.00 ± 0%      3.00 ± 0%     ~     (all equal)
SampleBuilderReordered-8           3.00 ± 0%      4.00 ± 0%  +33.33%  (p=0.029 n=4+4)
SampleBuilderFragmented-8          3.00 ± 0%      3.00 ± 0%     ~     (all equal)
SampleBuilderFragmentedLoss-8      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.029 n=4+4)
```